Replace unsigned int and int by NSUInteger and NSInteger

+++ bibdesk_vendorsrc/kenferry/KFASHandlerAdditions-TypeTranslation.m
@@ -26,9 +26,9 @@
 @end
 
 @interface NSNumber (KFAppleScriptHandlerAdditionsPrivate)
-+ (id) kfNumberWithSignedIntP:(void *)int_p byteCount:(unsigned int)bytes;
-+ (id) kfNumberWithUnsignedIntP:(void *)int_p byteCount:(unsigned int)bytes;
-+ (id) kfNumberWithFloatP:(void *)float_p byteCount:(unsigned int)bytes;
++ (id) kfNumberWithSignedIntP:(void *)int_p byteCount:(NSUInteger)bytes;
++ (id) kfNumberWithUnsignedIntP:(void *)int_p byteCount:(NSUInteger)bytes;
++ (id) kfNumberWithFloatP:(void *)float_p byteCount:(NSUInteger)bytes;
 @end
 
 @implementation NSObject (KFAppleScriptHandlerAdditions)
@@ -40,7 +40,7 @@
     if ([self respondsToSelector:@selector(objectEnumerator)]) 
     {
         id currentObject;
-        int i;
+        NSInteger i;
         
         resultDesc = [NSAppleEventDescriptor listDescriptor];
         NSEnumerator *objectEnumerator = [(id)self objectEnumerator];
@@ -81,7 +81,7 @@
 {
     NSAppleEventDescriptor *listDesc;
     NSMutableArray *resultArray;
-    int i, listCount;
+    NSInteger i, listCount;
     
     listDesc = [desc coerceToDescriptorType:typeAEList];
     resultArray = [NSMutableArray array];
@@ -103,7 +103,7 @@
     NSAppleEventDescriptor *resultDesc;
     NSMutableArray *userFields;
     NSArray *keys;
-    int keyCount, i;
+    NSInteger keyCount, i;
     
     resultDesc = [NSAppleEventDescriptor recordDescriptor];
     userFields = [NSMutableArray array];
@@ -143,7 +143,7 @@
 {
     NSMutableDictionary *resultDict;
     NSAppleEventDescriptor *recDescriptor, *listDescriptor;
-    int recordIndex, recordCount, listIndex, listCount, keyword;
+    NSInteger recordIndex, recordCount, listIndex, listCount, keyword;
     id keyObj, valObj;
         
     recDescriptor = [desc coerceToDescriptorType:typeAERecord];
@@ -249,7 +249,7 @@ static inline int areEqualEncodings(const char *enc1, const char *enc2)
 
 @implementation NSNumber (KFAppleScriptHandlerAdditions)
 
--(id)kfDescriptorValueWithFloatP:(void *)float_p byteCount:(unsigned int)bytes
+-(id)kfDescriptorValueWithFloatP:(void *)float_p byteCount:(NSUInteger)bytes
 {
     NSAppleEventDescriptor *resultDesc = nil;
     float floatVal;
@@ -286,7 +286,7 @@ static inline int areEqualEncodings(const char *enc1, const char *enc2)
     return resultDesc;
 }
 
--(id)kfDescriptorValueWithSignedIntP:(void *)int_p byteCount:(unsigned int)bytes
+-(id)kfDescriptorValueWithSignedIntP:(void *)int_p byteCount:(NSUInteger)bytes
 {
     NSAppleEventDescriptor *resultDesc;
     int intVal;
@@ -315,7 +315,7 @@ static inline int areEqualEncodings(const char *enc1, const char *enc2)
     return resultDesc;
 }
 
--(id)kfDescriptorValueWithUnsignedIntP:(void *)int_p byteCount:(unsigned int)bytes
+-(id)kfDescriptorValueWithUnsignedIntP:(void *)int_p byteCount:(NSUInteger)bytes
 {
     NSAppleEventDescriptor *resultDesc;
     unsigned int uIntVal;
@@ -510,7 +510,7 @@ static inline int areEqualEncodings(const char *enc1, const char *enc2)
     return resultNumber;
 }
 
-+ (id) kfNumberWithUnsignedIntP:(void *)int_p byteCount:(unsigned int)bytes
++ (id) kfNumberWithUnsignedIntP:(void *)int_p byteCount:(NSUInteger)bytes
 {
     NSNumber *resultNumber = nil;
     
@@ -543,7 +543,7 @@ static inline int areEqualEncodings(const char *enc1, const char *enc2)
     return resultNumber;
 }
 
-+ (id) kfNumberWithFloatP:(void *)float_p byteCount:(unsigned int)bytes
++ (id) kfNumberWithFloatP:(void *)float_p byteCount:(NSUInteger)bytes
 {
     NSNumber *resultNumber= nil;
